backend/requirements.txt
# ===== FASTAPI AND CORE DEPENDENCIES =====
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ===== INTERNET COMPUTER PROTOCOL =====
ic-py==2.5.0
candid==0.8.1
PyJWT==2.8.0

# ===== HTTP AND NETWORKING =====
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# ===== DATABASE =====
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0  # PostgreSQL async driver
aiosqlite==0.19.0  # SQLite async driver

# ===== SECURITY AND AUTHENTICATION =====
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==41.0.8

# ===== VALIDATION AND SERIALIZATION =====
email-validator==2.1.0
python-multipart==0.0.6

# ===== CACHING AND SESSIONS =====
redis==5.0.1
aioredis==2.0.1

# ===== LOGGING AND MONITORING =====
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# ===== FILE PROCESSING =====
python-magic==0.4.27
Pillow==10.1.0
PyPDF2==3.0.1

# ===== IPFS INTEGRATION =====
ipfshttpclient==0.8.0a2

# ===== DEVELOPMENT DEPENDENCIES =====
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# ===== DOCUMENTATION =====
mkdocs==1.5.3
mkdocs-material==9.4.8

# ===== ENVIRONMENT MANAGEMENT =====
python-dotenv==1.0.0

# ===== DATE AND TIME =====
python-dateutil==2.8.2

# ===== DATA PROCESSING =====
pandas==2.1.4
numpy==1.25.2

# ===== ASYNC UTILITIES =====
asyncio-throttle==1.0.2
aiolimiter==1.1.0

# ===== METRICS AND HEALTH CHECKS =====
prometheus-client==0.19.0
psutil==5.9.6

# ===== BLOCKCHAIN UTILITIES =====
coincurve==18.0.0
ecdsa==0.18.0

# ===== EMAIL NOTIFICATIONS =====
aiosmtplib==3.0.1
jinja2==3.1.2

# ===== CONFIGURATION =====
dynaconf==3.2.4

# ===== TESTING UTILITIES =====
factory-boy==3.3.0
faker==20.1.0
responses==0.24.1

# ===== PRODUCTION DEPENDENCIES =====
gunicorn==21.2.0
supervisor==4.2.5

# ===== SECURITY SCANNING =====
bandit==1.7.5
safety==2.3.4